requirements.txt
minio
reduct-py